Pagination and Pager Module

A pager module to add pagination to any collection of items. Given a number of items and a page break, the pagination module returns the indexes of the items on the currently selected page.

pager(id, n_items, page_break = 20)

paginationUI(id, width = 6, ..., offset = 3)

pagerUI(id, label_prev = "Previous", label_next = "Next",
  centered = TRUE, class = NULL)

pagerDemo(display.mode = c("showcase", "normal", "auto"))

Arguments

id
The shared id of the pagerUI() and the pager() module

n_items
The number of items in the collection of items to be paged

page_break
The maximum number of items in each page

width
The width of the containing column

...
Additional elements included in the containg column, added before the pagniation buttons

offset
The offest of the containing column

label_prev
The label of the "Previous" button.

label_next
The label of the "Next" button".

centered
If TRUE, the buttons are centered together, otherwise they are separated and spread to the left and right edges of their container.

class
Additional classes applied to the <nav> container.

display.mode
The mode in which to display the application. If set to the value "showcase", shows application code and metadata from a DESCRIPTION file in the application directory alongside the application. If set to "normal", displays the application normally. Defaults to "auto", which displays the application in the mode given in its DESCRIPTION file, if any.

Functions

paginationUI: The user interface for the pager module. The pager is contained in a single column in a fluidRow().

pagerUI: Next/previous buttons that can be used in addition to, or in place of paginationUI().

pagerDemo: Example app demonstrating usage of the pager module.

References

https://getbootstrap.com/docs/3.3/components/#pagination